backend/requirements.txt
fastapi>=0.115.0,<1.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-multipart>=0.0.9
pyjwt[crypto]>=2.9.0
supabase>=2.6.0
httpx>=0.27.0
instructor>=1.14.0,<2.0
openai>=1.40.0
anthropic>=0.40.0
jinja2>=3.1.0
